+++ crates/pvm/interpreter/src/memory.rs
@@ -5,41 +5,41 @@ use smallvec::SmallVec;
 use std::collections::BTreeMap;
 
 /// The size of a page in the memory.
-pub const PAGE_SIZE: u64 = 4096;
+pub const PAGE_SIZE: u32 = 4096;
 
 /// The memory of the interpreter.
 #[derive(Default, Debug, PartialEq, Eq, Clone)]
 pub struct Memory {
     /// The pages of the memory.
-    pub pages: BTreeMap<u64, Page>,
+    pub pages: BTreeMap<u32, Page>,
 }
 
 impl Memory {
     /// Read a value from the memory.
-    pub fn read<V: Value>(&self, address: u64) -> Result<V> {
+    pub fn read<V: Value>(&self, address: u32) -> Result<V> {
         self.read_offset(address, 0)
     }
 
     /// Read a value from the memory at an offset.
-    pub fn read_offset<V: Value>(&self, address: u64, offset: u64) -> Result<V> {
+    pub fn read_offset<V: Value>(&self, address: u32, offset: u32) -> Result<V> {
         let start = address + offset;
         let page = start / PAGE_SIZE;
         let offset = start % PAGE_SIZE;
 
         // read bytes
-        let bytes = self.read_bytes(page, offset, V::SIZE as u64)?;
-        V::from_bytes(&bytes).ok_or(Error::MemoryInaccessible(page as u32))
+        let bytes = self.read_bytes(page, offset, V::SIZE as u32)?;
+        V::from_bytes(&bytes).ok_or(Error::MemoryInaccessible(page))
     }
 
     /// Read bytes from the memory.
-    pub fn read_bytes(&self, page: u64, offset: u64, len: u64) -> Result<Vec<u8>> {
+    pub fn read_bytes(&self, page: u32, offset: u32, len: u32) -> Result<Vec<u8>> {
         if offset + len > PAGE_SIZE {
-            return Err(Error::MemoryInaccessible(page as u32));
+            return Err(Error::MemoryInaccessible(page));
         }
 
         let page = self.access(page)?;
         let data = page.data.as_slice();
-        let data_len = data.len() as u64;
+        let data_len = data.len() as u32;
 
         // fill with 0s if necessary
         let mut bytes = vec![0; len as usize];
@@ -50,16 +50,16 @@ impl Memory {
     }
 
     /// Write a value to the memory.
-    pub fn write<V: Value>(&mut self, address: u64, value: V) -> Result<()> {
+    pub fn write<V: Value>(&mut self, address: u32, value: V) -> Result<()> {
         self.write_bytes(address / PAGE_SIZE, address % PAGE_SIZE, &value.to_vec())
     }
 
     /// Write a value to the memory at an offset.
-    pub fn write_offset<V: Value>(&mut self, address: u64, offset: u64, value: V) -> Result<()> {
+    pub fn write_offset<V: Value>(&mut self, address: u32, offset: u32, value: V) -> Result<()> {
         let start = address + offset;
         let page = start / PAGE_SIZE;
         let offset = start % PAGE_SIZE;
-        if offset + V::SIZE as u64 > PAGE_SIZE {
+        if offset + V::SIZE as u32 > PAGE_SIZE {
             return Err(Error::MemoryInaccessible(page as u32));
         }
 
@@ -67,12 +67,12 @@ impl Memory {
     }
 
     /// Write bytes to the memory.
-    pub fn write_bytes(&mut self, page: u64, offset: u64, bytes: &[u8]) -> Result<()> {
+    pub fn write_bytes(&mut self, page: u32, offset: u32, bytes: &[u8]) -> Result<()> {
         let page = self.mutate(page)?;
 
         // extend page if necessary
-        let data_len = page.data.len() as u64;
-        let to_write = bytes.len() as u64;
+        let data_len = page.data.len() as u32;
+        let to_write = bytes.len() as u32;
         if data_len < to_write + offset {
             page.data.resize(to_write as usize + offset as usize, 0);
         }
@@ -84,7 +84,7 @@ impl Memory {
     }
 
     /// Convert the memory to a data map.
-    pub fn to_data_maps(&self) -> BTreeMap<u64, Vec<u8>> {
+    pub fn to_data_maps(&self) -> BTreeMap<u32, Vec<u8>> {
         let mut maps = BTreeMap::new();
 
         for (&page_num, page) in &self.pages {
@@ -106,7 +106,7 @@ impl Memory {
                     }
                 } else {
                     if current.is_none() {
-                        current = Some(base + offset as u64);
+                        current = Some(base + offset as u32);
                     }
                     data.push(byte);
                 }
@@ -124,14 +124,14 @@ impl Memory {
     }
 
     /// Get the access type of a memory slot.
-    fn access(&self, page: u64) -> Result<&Page> {
+    fn access(&self, page: u32) -> Result<&Page> {
         self.pages
             .get(&page)
             .ok_or(Error::MemoryInaccessible(page as u32))
     }
 
     /// Get the access type of a page.
-    fn mutate(&mut self, pagenum: u64) -> Result<&mut Page> {
+    fn mutate(&mut self, pagenum: u32) -> Result<&mut Page> {
         let page = self
             .pages
             .get_mut(&pagenum)
